test(community): re-add mark as read test in community tests
Fixes #9250
